+++ pkg/rpc/provider_events.go
@@ -1,0 +1,65 @@
+package rpc
+
+import (
+	"context"
+	"errors"
+
+	coretypes "github.com/cometbft/cometbft/rpc/core/types"
+)
+
+// Subscribe implements client.Client.
+// This functionality is currently not implemented.
+func (p *RpcProvider) Subscribe(ctx context.Context, subscriber string, query string, outCapacity ...int) (out <-chan coretypes.ResultEvent, err error) {
+	// Check if EventBus is available
+	if p.adapter.EventBus == nil {
+		return nil, errors.New("event bus is not configured, cannot subscribe to events")
+	}
+	// TODO: Implement subscription logic using p.adapter.EventBus
+	// Example structure (needs actual implementation):
+	// q, err := cmtquery.New(query)
+	// if err != nil {
+	// 	 return nil, fmt.Errorf("failed to parse query: %w", err)
+	// }
+	// sub, err := p.adapter.EventBus.Subscribe(ctx, subscriber, q, outCapacity...)
+	// if err != nil {
+	// 	  return nil, fmt.Errorf("failed to subscribe: %w", err)
+	// }
+	// // Need a way to convert EventBus messages to coretypes.ResultEvent
+	// outChan := make(chan coretypes.ResultEvent, ...) // Use appropriate capacity
+	// go func() {
+	// 	  for msg := range sub.Out() {
+	// 		  // Conversion logic here
+	// 		  outChan <- coretypes.ResultEvent{Query: query, Data: msg.Data(), Events: msg.Events()}
+	// 	  }
+	// 	  close(outChan)
+	// }()
+	// return outChan, nil
+	return nil, errors.New("event subscription functionality is not yet implemented")
+}
+
+// Unsubscribe implements client.Client.
+// This functionality is currently not implemented.
+func (p *RpcProvider) Unsubscribe(ctx context.Context, subscriber string, query string) error {
+	if p.adapter.EventBus == nil {
+		return errors.New("event bus is not configured, cannot unsubscribe")
+	}
+	// TODO: Implement unsubscription logic using p.adapter.EventBus
+	// Example structure:
+	// q, err := cmtquery.New(query)
+	// if err != nil {
+	// 	 return fmt.Errorf("failed to parse query: %w", err)
+	// }
+	// return p.adapter.EventBus.Unsubscribe(ctx, subscriber, q)
+	return errors.New("event unsubscription functionality is not yet implemented")
+}
+
+// UnsubscribeAll implements client.Client.
+// This functionality is currently not implemented.
+func (p *RpcProvider) UnsubscribeAll(ctx context.Context, subscriber string) error {
+	if p.adapter.EventBus == nil {
+		return errors.New("event bus is not configured, cannot unsubscribe")
+	}
+	// TODO: Implement unsubscribe all logic using p.adapter.EventBus
+	// return p.adapter.EventBus.UnsubscribeAll(ctx, subscriber)
+	return errors.New("event unsubscribe all functionality is not yet implemented")
+}
